Add event-driven pension projection bridge
