+++ full-monty.html
@@ -215,11 +215,11 @@
   </section>
 
   <script src="https://cdn.jsdelivr.net/npm/chart.js/dist/chart.umd.js"></script>
+
+  <!-- listener that handles fm-run-pension and emits fm-pension-output -->
+  <script type="module" src="./pensionProjection.js"></script>
+
   <script type="module" src="./fullMontyResults.js"></script>
   <script type="module" src="./fullMontyWizard.js"></script>
-  <script type="module">
-    import { openFullMontyWizard } from './fullMontyWizard.js';
-    openFullMontyWizard();
-  </script>
 </body>
 </html>
