+++ pension-projection.html
@@ -335,7 +335,7 @@
   <script src="https://cdn.jsdelivr.net/npm/jspdf-autotable@3.8.0/dist/jspdf.plugin.autotable.min.js"></script>
   <script type="module" src="pdfWarningHelpers.js"></script>
   <script src="./syncMinAge.js"></script>
-  <script type="module" src="./pensionProjection.js"></script>
+  <script type="module" src="./pensionProjectionPage.js"></script>
   <!-- Wizard overlay -->
   <div id="wizardModal" class="modal hidden">
     <div class="wizard-card">
